+++ test/integration/read_only_mode_integration_test.rb
@@ -9,7 +9,7 @@ class ReadOnlyModeIntegrationTest < ActionDispatch::IntegrationTest
   end
 
   test 'read only mode logs out' do
-    $stdout.stub(:puts, nil) do
+    disable_output do
       sign_in_as users(:user_regular)
       get root_path
       assert_select 'a#jupiter-user-nav-downdown', count: 1
@@ -22,7 +22,7 @@ class ReadOnlyModeIntegrationTest < ActionDispatch::IntegrationTest
 
   test 'read only mode creates and clears announcement' do
     skip('This consistently flaps for me locally')
-    $stdout.stub(:puts, nil) do
+    disable_output do
       sign_in_as users(:user_regular)
       get root_path
       assert_select 'div.alert', text: I18n.t('announcement_templates.read_only_mode'), count: 0
@@ -36,7 +36,7 @@ class ReadOnlyModeIntegrationTest < ActionDispatch::IntegrationTest
   end
 
   test 'cannot log in when read only mode enabled' do
-    $stdout.stub(:puts, nil) do
+    disable_output do
       get root_path
       assert_select 'a.nav-item', text: I18n.t('application.navbar.links.login'), count: 1
       Rake::Task['jupiter:enable_read_only_mode'].execute