+++ app/models/thesis.rb
@@ -17,8 +17,6 @@ class Thesis < JupiterCore::Doiable
   #
   # with an upgraded version of Postgresql this could be done more cleanly and performanetly
   scope :belongs_to_path, ->(path) { where('member_of_paths::text LIKE ?', "%#{path}%") }
-  scope :updated_on_or_after, ->(date) { where('updated_at >= ?', date) }
-  scope :updated_before, ->(date) { where('updated_at < ?', date) }
 
   before_validation :populate_sort_year
   after_save :push_entity_for_preservation
@@ -53,6 +51,13 @@ class Thesis < JupiterCore::Doiable
   def add_to_path(community_id, collection_id)
     self.member_of_paths ||= []
     self.member_of_paths += ["#{community_id}/#{collection_id}"]
+  end
+
+  def ordered_files
+    # We are sorting with lowercase filenames so we get a list that would be
+    # more familiar to end users mixing upper and lower case in the final order
+    # like 0-9,a-z
+    files.joins(:blob).order('LOWER(active_storage_blobs.filename) ASC')
   end
 
   def self.from_draft(draft_thesis)