+++ CHANGELOG.md
@@ -11,7 +11,7 @@ New entries in this file should aim to provide a meaningful amount of informatio
 
 ### Added
 - Push collections and communities to preservation on save along with a rake task to do so [#255](https://github.com/ualbertalib/pushmi_pullyu/issues/255)
-
+- Added minitest-retry gem to retry flapping tests that are able to pass through retries [#3044](https://github.com/ualbertalib/jupiter/pull/3044)
 ## [2.4.3] - 2022-12-14
 
 ### Security